+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the "php-docblocker" extension will be documented in this
 
 ## [Unreleased]
 - Add support for params which have a trailing comma
+- Add support for constructor promotion
 - Add GitHub workflow to prepare and publish releases
 
 ## [2.3.0] - 2021-10-23